Remove circular dependency when applying modules middleware loader
(When I say middleware loader here, I mean the thing that powers
`--test-scheduled`, not Wrangler's facade mechanism which D1 uses)

For modules mode, the middleware loader is split into two esbuild
stages. The first "insertion" stage generates a file that imports the
current entrypoint (either the user's worker or output of another
facade, e.g. D1) and all the selected middleware. The second "loader"
stage re-exports this output using
https://github.com/cloudflare/workers-sdk/blob/ff35b9d1f589e2d07e0d1935a1c53b90de01588a/packages/wrangler/templates/middleware/loader-modules.ts.
Note all imports are marked as external in these stages, so no
bundling occurs. Once all Wrangler facades are applied (not just the
middleware loader), we run esbuild a final time to produce a Worker
bundle.

Prior to this change, the second "loader" stage would output to the
same location as the input to the first "insertion" stage, meaning we
ended up with a circular import. See the PR description for a full
example.

This PR updates the "loader" stage to output to a new file,
preventing the circular dependency.

Alternative to #2682

Closes #2666
Closes #2054